fix(TUP-23672):improve for check connection button for tacokit integration to Studio
